Refactor connection usages
