+++ middleware/middleware.go
@@ -2,8 +2,6 @@ package middleware
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"github.com/sirupsen/logrus"
 	"os"
 	"time"
@@ -52,22 +50,6 @@ func GetLoggerFile() *os.File {
 	return openLogfile
 }
 
-func HandleMapMarshal(data map[string]interface{}) ([]byte, error) {
-	jsonByteData, err := json.Marshal(data)
-	if err != nil {
-		return nil, errors.New("can't parse map object")
-	}
-	return jsonByteData, err
-}
-
-func HandleMapUnmarshal(data string, mp *map[string]interface{}) error {
-	err := json.Unmarshal([]byte(data), &mp)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 func Initialize(ctx context.Context, username string, password string) error {
 	if clientConfig == nil {
 		clientConfig = &config{}
@@ -102,76 +84,6 @@ func IsClientInitialized() bool {
 		return true
 	}
 	return false
-}
-
-func CanVerifyVolumes() (bool, error) {
-	if !(IsClientInitialized()) {
-		return clientConfig.verifyVolumes, errors.New("middleware could not be initialized")
-	}
-	return clientConfig.verifyVolumes, nil
-}
-
-func CanVerifyAttachPath() bool {
-	return clientConfig.verifyAttachedPath
-}
-
-func CanVerifyLockedVolumes() bool {
-	return clientConfig.verifyLockedPath
-}
-
-func GetIgnorePaths() []string {
-	return clientConfig.ignorePaths
-}
-
-func GetRootDataset() string {
-	return clientConfig.appsDataset
-}
-
-func generateSocket(ctx context.Context, socketUrl string, username string, password string) error {
-	conn, _, connErr := websocket.Dial(ctx, socketUrl, nil)
-	if connErr != nil {
-		return connErr
-	}
-	conn.SetReadLimit(32769 * 10)
-	connectionResp, connErr := GenerateSession(ctx, conn)
-	if connErr != nil {
-		return connErr
-	}
-	if (username != "") && (password != "") {
-		_, loginErr := LoginSession(ctx, conn, connectionResp["session"].(string), username, password)
-		if loginErr != nil {
-			return loginErr
-		}
-	}
-	clientConfig.client = &Client{
-		id:       connectionResp["session"].(string),
-		msg:      "method",
-		ctx:      ctx,
-		conn:     conn,
-		username: username,
-		password: password,
-	}
-	return nil
-}
-
-func Call(method string, params ...interface{}) (map[string]interface{}, error) {
-	m := clientConfig.client
-	resp, err := m.get(method, params...)
-	return resp, err
-}
-
-func (m *Client) get(method string, params ...interface{}) (map[string]interface{}, error) {
-	if m == nil {
-		return nil, errors.New("client is not initialized")
-	}
-	data := map[string]interface{}{
-		"id":     m.id,
-		"msg":    m.msg,
-		"method": method,
-		"params": params,
-	}
-	connResp, connErr := handleSocketCommunication(m.ctx, m.conn, data)
-	return connResp, connErr
 }
 
 func (m *Client) Close() {
